+++ cmd/root.go
@@ -7,7 +7,7 @@ import (
 	"os"
 )
 
-const version = "1.11.0"
+const version = "1.12.0"
 
 // Execute is the entry point called by main.
 func Execute() int {